pyproject.toml
[tool.poetry]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = ["alstacon <alstacon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "2.2"
djangorestframework = "3.13"
django-environ = "^0.10.0"
psycopg2-binary = "2.8.6"
django-filter = "21.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
